requirements.txt
python-fasthtml
monsterui
claudette
